Updates for atoi, bzero, isalnum, memcpy and memset.

+++ ft_memcpy.c
@@ -14,9 +14,9 @@
 
 void	*ft_memcpy(void *dest, const void *src, size_t n)
 {
-	char			*dest_char_ptr;
-	const char		*src_char_ptr;
-	unsigned int	i;
+	char		*dest_char_ptr;
+	const char	*src_char_ptr;
+	size_t		i;
 
 	dest_char_ptr = dest;
 	src_char_ptr = src;
